ODC-7655: Migrate typeahead `SelectInputField` to PF5
Remove `SelectInputField`
