+++ frontend/packages/console-shared/src/components/formik-fields/field-types.ts
@@ -1,5 +1,4 @@
 import { ValidatedOptions, TextInputTypes, gridItemSpanValueShape } from '@patternfly/react-core';
-import { SelectVariant as SelectVariantDeprecated } from '@patternfly/react-core/deprecated';
 import { JSONSchema7 } from 'json-schema';
 import { K8sKind, K8sResourceKind } from '@console/internal/module/k8s';
 import { RowRendererProps } from './multi-column-field/MultiColumnFieldRow';
@@ -175,42 +174,29 @@ export interface SelectInputOption {
   hasCheckbox?: boolean;
 }
 
-export interface SingleDropdownFieldProps extends FieldProps {
-  ariaLabel?: string;
-  options: SelectInputOption[];
-  placeholderText?: React.ReactNode;
-  isDisabled?: boolean;
-  toggleOnSelection?: boolean;
-  onChange?: (selection: string) => void;
-  getLabelFromValue?: (value: string) => string;
-}
-
 export interface SelectInputFieldProps extends FieldProps {
   ariaLabel?: string;
   options: SelectInputOption[];
-  variant?: SelectVariantDeprecated;
   placeholderText?: React.ReactNode;
-  isCreatable?: boolean;
   isDisabled?: boolean;
-  hasOnCreateOption?: boolean;
-  isInputValuePersisted?: boolean;
-  noResultsFoundText?: string;
   toggleOnSelection?: boolean;
-  hideClearButton?: boolean;
   onChange?: (selection: string) => void;
+}
+
+export interface SingleDropdownFieldProps extends SelectInputFieldProps {
   getLabelFromValue?: (value: string) => string;
 }
 
-export interface MultiTypeaheadFieldProps extends FieldProps {
-  ariaLabel?: string;
-  options: SelectInputOption[];
-  placeholderText?: string;
+export interface MultiTypeaheadFieldProps extends SelectInputFieldProps {
   isCreatable?: boolean;
-  isDisabled?: boolean;
   isInputValuePersisted?: boolean;
   noResultsFoundText?: string;
-  toggleOnSelection?: boolean;
   hideClearButton?: boolean;
-  onChange?: (selection: string) => void;
   getLabelFromValue?: (value: string) => string;
+}
+
+export interface SelectTypeaheadDropdownProps extends SelectInputFieldProps {
+  hasOnCreateOption?: boolean;
+  hideClearButton?: boolean;
+  isCreatable?: boolean;
 }
