+++ frontend/packages/console-shared/src/components/formik-fields/index.ts
@@ -20,9 +20,9 @@ export * from './text-column-field/text-column-types';
 export { default as TextColumnField } from './text-column-field/TextColumnField';
 export { default as DynamicFormField } from './DynamicFormField';
 export { default as SyncedEditorField } from './SyncedEditorField';
-export { default as SelectInputField } from './SelectInputField';
 export { default as SingleDropdownField } from './SingleDropdownField';
 export { default as MultiTypeaheadField } from './MultiTypeaheadField';
+export { default as SingleTypeaheadField } from './SingleTypeaheadField';
 export { default as SelectorInputField } from './SelectorInputField';
 export * from './field-utils';
 export * from './field-types';
