+++ frontend/packages/dev-console/locales/en/devconsole.json
@@ -525,7 +525,6 @@
   "Path that the router watches to route traffic to the service.": "Path that the router watches to route traffic to the service.",
   "Target port": "Target port",
   "Target port for traffic.": "Target port for traffic.",
-  "No results found": "No results found",
   "Edge": "Edge",
   "Passthrough": "Passthrough",
   "Re-encrypt": "Re-encrypt",
